Make size/position types generic over pixel type (#1277)
* Begin implementing DPI generics

* Fix multithreaded example

* Format

* Fix serde test

* hopefully fix most of the errors

* Fix dpi module errors

* More error fixings

* Format

* fix macos errors

* Another error pass

* Replace bad type signatures

* more fixins

+++ src/platform_impl/linux/x11/event_processor.rs
@@ -704,7 +704,7 @@ impl<T: 'static> EventProcessor<T> {
                         });
                         if cursor_moved == Some(true) {
                             let position =
-                                PhysicalPosition::new(xev.event_x as f64, xev.event_y as f64);
+                                PhysicalPosition::new(xev.event_x as i32, xev.event_y as i32);
 
                             callback(Event::WindowEvent {
                                 window_id,
@@ -809,7 +809,7 @@ impl<T: 'static> EventProcessor<T> {
                         });
 
                         let position =
-                            PhysicalPosition::new(xev.event_x as f64, xev.event_y as f64);
+                            PhysicalPosition::new(xev.event_x as i32, xev.event_y as i32);
 
                         // The mods field on this event isn't actually populated, so query the
                         // pointer device. In the future, we can likely remove this round-trip by
@@ -888,7 +888,7 @@ impl<T: 'static> EventProcessor<T> {
                             .unwrap_or(2);
 
                         let position =
-                            PhysicalPosition::new(xev.event_x as f64, xev.event_y as f64);
+                            PhysicalPosition::new(xev.event_x as i32, xev.event_y as i32);
 
                         callback(Event::WindowEvent {
                             window_id,
